replace the default numa01 with numa01_NO_BIND_FORCE_SAME_NODE
With numa01 the kernel MPOL_DEFAULT doesn't make it start from the
worst possible memory layout configuration. The worst layout is with
the memory of both processes in the same NUMA node. It's not worth
running both numa01 and numa01_NO_BIND_FORCE_SAME_NODE versions
because the result from numa01_NO_BIND_FORCE_SAME_NODE is the more
interesting of the two. numa01_NO_BIND_FORCE_SAME_NODE exacerbates to
the maximum the NUMA effects and it is also much harder for any
automatic NUMA placement algorithm to solve.

numa01_NO_BIND_FORCE_SAME_NODE allocates the RAM of both processes in
the same node, so AutoNUMA must make an arbitrary random decision (one
that wouldn't need to make with numa01 default because a little ram of
one of the two process would spill over to the other node and the
weights wouldn't be exactly equal for both processes).

Because at startup the NUMA weights are exactly equal for both
processes with _NO_BIND_FORCE_SAME_NODE, what happens is that the
last_nid anti false sharing logic will eventually trigger more false
positives in one process than the other. Pretty quickly one of the two
processes will have a little more memory spillover than the other
process in the previously empty node. Statistically the one with more
spillovers will also be the process that has more threads running in
the previously empty node (which is a positive as we'll have to move
less threads across CPU too). From that point AutoNUMA knows exactly
how to converge and one process will be moved over completely to the
previously empty node, and the other process will get that little
memory spillover that was moved to the previously empty node, back to
its original node. The spillover for the process that converges in the
original node is superfluous but that kind of spillover randomness is
needed to move away from the initial status that AutoNUMA can't take
an decision. Some randomness is needed to solve the
numa01_NO_BIND_FORCE_SAME_NODE problem, and the false positives
resulting from the last_nid anti-false-sharing logic provides for it.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>